weapon class base 생성 및 cpp 코드 폴더 분할

+++ Source/Hola/InteractObject/InteractObject.cpp
@@ -40,7 +40,7 @@ void AInteractObject::Tick(float DeltaTime)
 
 }
 
-float AInteractObject::GetNeedEnergy()
+int AInteractObject::GetNeedEnergy()
 {
 	return needEnergy;
 }
